fix(앱): 반경을 소수로 받는다 — 자동 반경이면 폰이 서버 응답을 통째로 버렸다 (#115 · 2.9.4-radiusdouble)
기사님 «7지점 한 바퀴» — 반경을 «40km 기준(자동)»으로 켜자 콜이 하나도 안 잡혔다 (2026-09-14 14:41)
- 서버가 pickupRadiusKm: 4.554354460578365 를 보냈고, 앱 FilterConfig 는 Int 라 Gson 이 응답 전체를 버렸다
  (NumberFormatException · 14:37:33 부터 매번). 새 필터·기기 모드·판결을 같이 못 받아
  접근성을 다시 켠 뒤 모드가 기본값 MANUAL 로 남았고, 필터를 통과한 콜을 누르지 않았다
- 서버는 09-12 cb82c42 부터 자동 반경을 소수로 보냈다. appFilterKeys 는 이름만 대조해 초록이었다

고침
- SharedModels.kt FilterConfig: pickupRadiusKm · destinationRadiusKm → Double
- 인성·화물24시·픽커 파서: optInt → optDouble (4.55 를 4 로 자르던 것) · 픽커 decide 인자 Double
- 앱 단위 검사의 정수 인자 → .0
- 앱 버전 2.9.4-radiusdouble (52) — 폰에 설치함

검사 (먼저 빨간불)
- appFilterKeys.test.ts «서버가 소수로 보낼 수 있는 칸을 앱이 정수로 받지 않는다»
  소수가 되는 칸은 서버가 쓰는 effectiveRadii 에 자동을 켜서 뽑는다 (손 목록 아님)
  빨간불: pickupRadiusKm · destinationRadiusKm → 초록

게이트: jest 166/166 · 앱 assembleDebug · testDebugUnitTest · audit:docs
버그 대장 #115 — 남은 것: 앱이 한 칸만 틀려도 응답 전체를 버리는 구조

+++ onedal-app/app/src/main/java/com/onedal/app/models/SharedModels.kt
@@ -346,7 +346,11 @@ data class FilterConfig(
     val isActive: Boolean = false,
     val isSharedMode: Boolean = false,
     // ── 이하 기본값은 서버 미응답 시 최후 안전망 (서버 기본값과 같은 값으로 맞춘다) ──
-    val pickupRadiusKm: Int = 10,
+    /**
+     * 🔴 **소수로 받는다** (2026-09-14 · `appFilterKeys.test.ts`) — 자동 반경이면 서버가 `4.55` 처럼 보낸다.
+     *    `Int` 였을 때 Gson 이 응답을 **통째로 버려** 새 필터도 모드도 못 받고 MANUAL 로 남았다.
+     */
+    val pickupRadiusKm: Double = 10.0,
     /** 서버 기본값과 동일. 0 으로 두면 하한이 사라져 아무 콜이나 잡는다 */
     val minFare: Int = 30000,
     /** 🔔 픽커 알람 요금 하한 — 원천 DB(user_settings.picker_alarm_min_fare) → 피기백 (2026-08-30) */
@@ -359,7 +363,7 @@ data class FilterConfig(
      */
     val ratePerKm: Map<String, Int> = emptyMap(),
     val destinationCity: String = "",
-    val destinationRadiusKm: Int = 10,
+    val destinationRadiusKm: Double = 10.0,   // 🔴 소수로 받는다 — 위 pickupRadiusKm 과 같은 까닭
     val excludedKeywords: List<String> = emptyList(),
     val destinationKeywords: List<String> = emptyList(),
     val customCityFilters: List<String> = emptyList(),
